+++ src/ios/Launcher/LauncherViewModel.swift
@@ -18,10 +18,6 @@ class LauncherViewModel: NSObject, ObservableObject {
   
   @Published var multiplayerConfig: MultiplayerConfig?
   
-  #if ZERO
-  @Published var isFullVersionPurchased: Bool = false
-  #endif
-  
   var currentConfig: LauncherConfig? {
     didSet {
       if let currentConfig {